+++ haowugou-infrastructure/src/main/java/com/haowugou/infrastructure/persistence/sales/StoreDailySalesDataObject.java
@@ -1,4 +1,4 @@
-package com.haowugou.infrastructure.persistence.data;
+package com.haowugou.infrastructure.persistence.sales;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;